ELF symbols: Detect real vaddr when needed (ARM Thumb)
Turns out that on ARM 32-bit ELF symbols referring to functions whose
instructions are encoded in Thumb mode can have misaligned virtual addresses
(i.e. not multiples of 2). This is both good and bad:

- Good because the entries in the syscall table will also be misaligned
  addresses (kernel code needs to switch mode calling those functions), so it's
  easy to match them with the right symbol.
- Bad because if we need to refer to the actual function (e.g. to invoke
  addr2line or to read its code) we need the real (aligned) address.

Add a .real_vaddr field to the Symbol class and assign it a value different
than .vaddr when needed.