arch, kernel: Support extracting more esoteric syscalls
Some (insane) esoteric syscalls are implemented in weird ways that make binary
analysis the only option to detect them. For example: PowerPC's switch_endian
has a fast implementation with number 0x1ebe that is implemented in asm at the
syscall entry point. We need an actual function to extract these.